Make all tests pass
