+++ debug/test/browser/debug.test.js
@@ -15,10 +15,6 @@ import {
 import './fakeDevTools';
 import 'preact/debug';
 import { setupRerender } from 'preact/test-utils';
-import * as PropTypes from 'prop-types';
-
-// eslint-disable-next-line no-duplicate-imports
-import { resetPropWarnings } from 'preact/debug';
 
 const h = createElement;
 /** @jsx createElement */
@@ -790,91 +786,6 @@ describe('debug', () => {
 			);
 
 			render(<Anchor />, scratch);
-			expect(console.error).to.not.be.called;
-		});
-	});
-
-	describe('PropTypes', () => {
-		beforeEach(() => {
-			resetPropWarnings();
-		});
-
-		it("should fail if props don't match prop-types", () => {
-			function Foo(props) {
-				return <h1>{props.text}</h1>;
-			}
-
-			Foo.propTypes = {
-				text: PropTypes.string.isRequired
-			};
-
-			render(<Foo text={123} />, scratch);
-
-			expect(console.error).to.be.calledOnce;
-
-			// The message here may change when the "prop-types" library is updated,
-			// but we check it exactly to make sure all parameters were supplied
-			// correctly.
-			console.log(console.error.args);
-			expect(console.error).to.have.been.calledOnceWith(
-				sinon.match(
-					/^Failed prop type: Invalid prop `text` of type `number` supplied to `Foo`, expected `string`\.\n {2}in Foo \(at (.*)[/\\]debug[/\\]test[/\\]browser[/\\]debug\.test\.js:[0-9]+\)$/m
-				)
-			);
-		});
-
-		it('should only log a given prop type error once', () => {
-			function Foo(props) {
-				return <h1>{props.text}</h1>;
-			}
-
-			Foo.propTypes = {
-				text: PropTypes.string.isRequired,
-				count: PropTypes.number
-			};
-
-			// Trigger the same error twice. The error should only be logged
-			// once.
-			render(<Foo text={123} />, scratch);
-			render(<Foo text={123} />, scratch);
-
-			expect(console.error).to.be.calledOnce;
-
-			// Trigger a different error. This should result in a new log
-			// message.
-			console.error.resetHistory();
-			render(<Foo text="ok" count="123" />, scratch);
-			expect(console.error).to.be.calledOnce;
-		});
-
-		it('should render with error logged when validator gets signal and throws exception', () => {
-			function Baz(props) {
-				return <h1>{props.unhappy}</h1>;
-			}
-
-			Baz.propTypes = {
-				unhappy: function alwaysThrows(obj, key) {
-					if (obj[key] === 'signal') throw Error('got prop');
-				}
-			};
-
-			render(<Baz unhappy={'signal'} />, scratch);
-
-			expect(console.error).to.be.calledOnce;
-			expect(errors[0].includes('got prop')).to.equal(true);
-			expect(serializeHtml(scratch)).to.equal('<h1>signal</h1>');
-		});
-
-		it('should not print to console when types are correct', () => {
-			function Bar(props) {
-				return <h1>{props.text}</h1>;
-			}
-
-			Bar.propTypes = {
-				text: PropTypes.string.isRequired
-			};
-
-			render(<Bar text="foo" />, scratch);
 			expect(console.error).to.not.be.called;
 		});
 	});
